Performance optimization for scrolling sharedElements
This change improves the performance for shared element
modifiers during scrolling by avoiding eagerly tracking
position changes for dormant shared elements. The
position is instead obtained just in time before
starting a bounds transform animation.

With this change, the amount of placements on layout nodes
are the same with shared element modifiers vs. without. This
optimization significantly benefits lazy lists where each
item has a shared element/bounds modifier on it, which is
not uncommon in the apps.

This change also fixes an issue in the skipToLookahead
placement calculation. New test is added to prevent regression.

Test: New tests added

Change-Id: I93dadd8779404a3151c77bb3fcedfe3545a033d6