Revert^2 "Add window insets rulers"
Bug: 406258408
Bug: 359260964
Fixes: 405456048
Bug: 406258408

This reverts commit 8aae5a3f53aae8b44643f4787a109480775251c5.

Reason for revert: to fix problems with performance and g3

Relnote: "Adds Rulers for Window Insets. Adds DerivedRuler
to allow a Ruler to be calculated from another Ruler.
Modifies PlacmentScope to implement Density."

Rulers allow layouts to align to specific absolute
elements. WindowInsets are a perfect use case for
Rulers.

Test: new tests, manually ran performance and g3 tests

Change-Id: I27f22bd4abdf1c09ef8a7fda4bd4ebb68fa02adc