Fix regression: Voice access can't find SlidingPaneLayout divider
This CL reverted the changes aosp/3529082, and use the different
way to prevent unfocused SlidingPaneLayout divider being scrolled by
Talkback gestures.

Talkback scroll gestures are sent to all scrollable elements despite
that they are not focused. In the previous CL, we make the divider not
scrollable when it's unfocused. However, the Voice Access will ignore
the divider because it doesn't have any actions.

This CL takes another approach, and divider is always scrollable but it
won't accept scroll action if it's not focused. So that Voice Access can
correctly find the divider as well.

Bug: 378798918
Test: SlidingPaneLayoutA11yTest
Change-Id: I1e880453c9843b78e86de42e23a2bf8202fd8317