Make TimePickerBenchmark test resolution independent
Before the change, the test assumed that hour picker was always located at x = 1/4 of the screen. However, that assumption doesn't hold for higher resolutions.
As a result of that assumption, hourX and amPmX swipings were outside of TimePicker, making this benchmark behave differently on different resolutions.

Bug: 405342865
Relnote: N/A
Test: manual verification that TimePickerBenchmarkTest swipes correctly on different devices

Change-Id: I3863af3e512812c3d44b1643f17541b448ac7c33