Expose the SessionConfig APIs
Added the new SessionConfig APIs to allow apps to configure
the camera session parameters that applies to the camera as
well as the CameraEffect and ViewPort.

To use the new APIs,
apps can create the SessionConfig instance with the use cases
with or without CameraEffect and ViewPort. And bind the
sessionConfig to the lifecycleOwner.

For example:
cameraProvider(lifecycleOwner, cameraSelector,
      SessionConfig(
         useCases = listOf(preview, imageCapture)
         effects = listOf(effect),
         viewPort = viewport
      )
)

doc:
go/CameraX-sessionconfig

Bug: 414298514
Test: ProcessCameraProviderTest / LifecycleCameraTest
RelNote: "Introduced the new SessionConfig API, which includes
a better abstraction for configuring use cases,
session parameters, CameraEffect, and ViewPort and etc.
Binding a SessionConfig to a LifecycleOwner opens the camera
session, configures it using the specified use cases and
session parameters, and applies the designated CameraEffect
and ViewPort. When updating a new SessionConfig to the
same LifecycleOwner, you can just bind a new Sessionconfig
without the need of invoking unbind or unbindAll first."

Change-Id: Iedfc38a69cd6d67e2791381700545ebd2d4ba4ca