Adds @RememberInComposition
RememberInComposition is an annotation that can mark constructors, functions, and property getters, to indicate that they must not be called directly inside composition, without being remembered. This is enforced by a corresponding lint check.

Bug: b/381406389
Test: RememberInCompositionDetectorTest
Relnote: "Adds @RememberInComposition: an annotation that can mark constructors, functions, and property getters, to indicate that they must not be called directly inside composition, without being remembered. Errors will be raised by a corresponding lint check. This annotation can be used to mark declarations that return stateful / mutable objects, objects whose identity is important to maintain across compositions, or objects that are expensive to instantiate and should be cached across compositions."
Change-Id: Ie7db088c6d601b39d7f0bbb5f877f961108b6c47