Adds manual Compose integration testing surface
This is a new Fragment launched by a new scenario button in the test
app. There is no new Activity.

It is impossible to leverage PdfViewerFragment for this as PVF expects
specific View classes to be present. It might theoretically be possible
to leverage PdfDocumentViewModel for this Fragment, but 1) that would
require opening up that (non-API) class to at least the library group
2) I would like to build this integration testing surface from the
ground up in a Compose-first way 3) PdfDocumentViewModel exposes events
and state that won't immediately be relevant to the Compose
implementation.

Test: ./gradlew :pdf:integration-tests:testapp:(test|cAT), ./gradlew :pdf:pdf-compose:test, manual
Bug: 393195927
Change-Id: I45036980bc27a34dfb3613f7a8d5e66f8aab6a09