Prepare for pausedPrecomposition switching isComplete from true to false between frames in prefetch
Right now it is not practically happening because of b/407931790, but after aosp/3571365 it will start happening, and it was always intended to work this way.
We shouldn't cache isComposed value, instead we should always check pausedPrecomposition.isComplete before doing apply, even if we though we completed the composition in the last frame.

Test: existing tests, manually with logs
Bug: 407931790
Change-Id: I34c20aab5f84dc91b40d3e87d7eacdcf11486f4d