Support RoomDatabase.runInTransaction with SQLiteDriver
Add support for blocking runInTransaction() API that is a RoomDatabase member. Even though this is technically a SupportSQLite API that is not supported once a SQLiteDriver is installed, for better compatibility and migration we can update its implementation to perform the given body in a transaction form the connection pool by bridging it with runBlocking.

Also add @throws documentation to various other RoomDatabase APIs that are not supported when a SQLiteDriver is installed.

Bug: 408364828
Test: InTransactionTest
Change-Id: I6dad52b6432cd2e801572f0d7d5f43774109efe1