TextLayoutState.getOffsetForPosition returning negative value
Not placing `innerTextField` in `TextFieldDecorator` leads to a lack of `TextLayoutResult` in `TextLayoutState`. In this case `TextLayoutState.getOffsetForPosition` returns `-1`. However some callers pass this offset to a `TextRange` in their downstream, causing a crash since `TextRange` does not accept negative values.

Here we address the few remaining flows that pass the returned value from `getOffsetForPosition` to a `TextRange` without a check.

Fix: b/308398612
Test: ./gradlew :compose:foundation:foundation:cAT
Change-Id: I2638c2cffea1ac5f4bd8ce105cdfc79a2503f0ab