Update ink experimental properties
These projects targets both Java and Kotlin clients. Experimental properties are no longer allowed in projects with Java clients because the lint check which enforces experimental usage for Java clients is not able to detect that accessors of experimental properties are experimental, and the kotlin compiler does not allow experimental annotations directly on property accessor targets. Previously this compiler error was suppressed, but that will no longer be allowed in Kotlin 2.0.

This makes the existing experimental properties private, and adds experimental accessor functions.

Bug: 409363281
Test: `./gradlew checkApi`, `./gradlew lint`
Relnote: Previously experimental property `InProgressStrokesView.textureBitmapStore` is now private. There are still public accessors for this property.
Change-Id: I1d7064ff25c953066a4b43a6c30a4b3aa0b145d5