Rename appfunctions-common to appfunctions
This is part of the package restructuring process for AppFunctions
Jetpack. The goal is to make a layered structure for developers to start
from any layer depends on their requirements. And also make sure the
package name matches the module name.

appfunctions module would be the module provides all the APIs to use
AppFunctions. For example, an assistant app can take appfunctions module
to write the function calling logic. Or an app that wants to write their
own runtime infrastructure to expose AppFunctions can use this module
that provide Compat APIs.

Relnote: N/A
Bug: 408453192
Test: Integration tests
Change-Id: I4d2e85e1f0e7afdff7e64459dd261f99030513e4