Use AbstractKotlinCompile to encompass all targets and set yarn vars using YarnRootEnvSpec
In AndroidXComposeImplPlugin,Kotlin2JsCompile tasks did not have the compose compiler added to them as they are not a super-class of KotlinCompile tasks.
This became evident in the child change, where WASM tests were failing as they needed the compose plugin.

Changing to AbstractKotlinCompile ensures we capture Kotlin2JsCompile type tasks as well as others.

We also update setting yarn variables using the properties available in YarnRootEnvSpec

BUG: 375595154
Change-Id: I19fe04b1755877f45f9e2aa978cab4d8da099b7a