Migrate appfunctions-common/schema/runtime to Kotlin 2.0
There is an issue with isVal with current KSP2 version that has been
fixed with the latest version. However, we should not use isVal to check
if the parameter has getter, since it's possible have var in primary
constructor as well.

Therefore, inspect the property declaration to check if the parameter
has a corresponding getter instead of relying on isVal.

Bug: 407634710
Test: Integration tests
Change-Id: If95ec903e27162edb574eae8f6c94ad4c1bd04c3