Change entry used for animations
Previously, SceneNavDisplay would use animations from the incoming entry. Now it uses the animations from the entry that has the highest z-index. This is done by taking the last entry of the scene with the highest z-index.

This means that when navigating, it uses the transition animation from the incoming scene/entry. When popping or during predictive back, it uses the popTransition animations from the outgoing scene/entry.

Bug: 409630339
Test: ./gradlew navigation3:navigation3:cC
Relnote: "SceneNavDisplay now uses the animation from the top-most rendered entry. This means that when navigating, it uses the transition animation from the incoming entry. When popping or during predictive back, it uses the popTransition animations from the outgoing entry."
Change-Id: Icab0eab2e960b23ac7f156a4ad1c2de77b070e2b