Reduce test to fit within 30 minutes
Fixes: 409858150

testCosSinError() previously checked all float values between
0 and 360 using integers and toRawBits() and  fromRawBits().
This could be a problem for two reasons: if there are too many
float values to complete a test within the allotted time or
the backing integer doesn't conform to incrementing correctly
between the two numbers.

In fact 0 is represented as all zero values, and incrementing
2^23 times gets to 10^-38. Incrementing to 1 is 2^30. I can imagine
it can take a long time to execute that code.

I've made the test run from 0 - 360, incrementing by 0.01 so
that it is relatively few iterations.

Test: ran locally
Change-Id: I65ee26171af0b5e868ea480007e69307d9c1d436