Fix the usage of coroutine scope in adaptive samples
Since panes can be shown or hidden, the coroutine scope remembered in a
pane will be forgotten when the pane is hidden. This triggers an issue
of SeekableTransition - it waits for composition to be done via waiting
a CancellableContinuation to finish before invoking onTransitionEnd(),
however once the coroutine scope is forgotten, the continuation instance
will get cancelled, which causes onTransitionEnd() is never run and
inconsistent transition states.

Fixes the issue on our end by remembering the coroutine scope at the
scaffold level to ensure it won't be forgotten during scaffold's
life-cycle. Will file a bug on the SeekableTransition to fix the root
cause as well.

Test: manual
Change-Id: I57208027db1573864c933f2980f3b855237dd32e