Cancel BroadcastFrameClock awaiters without locks
In the referenced bug, there's a deadlock where
a call to `withFrameNanos` is being canceled on one thread while another
thread is dispatching another thread. To avoid the deadlock, this
commit updates the BroadcastFrameClock awaiter so that it's possible
to cancel an awaiter without acquiring any locks.

Fixes: b/407027032
Test: BroadcastFrameClockTest.locklessCancellation
Relnote: "Fixed a deadlock that may affect Molecule users when a
suspended call to `FrameClock.withFrameNanos` is cancelled while a
frame is being dispatched."

Change-Id: I89cab8e3eab14ed9a85b36e151f11b5f526a01fd