Fix the issue that androidx.camera.media3.effect causes runtime crash when using with media3 1.6 dependency.
androidx.camera.media3:media3-effect uses media3 1.5 depdendency
which is not compatibile with media3 1.6.

Solution: update the dependency to 1.6 and correct the API usage
and also fix the issue that preview stops working
after the setting the effect more than once.

Bug: 407788366
Bug: 410482100
Test: Media3EffectDeviceTest
RelNote: "Fixed the cameraX media3-effect crash when using
with media3 1.6 dependency"

Change-Id: Ic1ff129dc4467fc7fa376de3497a941231415fda