Address TooManyRequestsException due to registering too many NetworkCallback.
This changes updates the required NetworkRequest tracker to avoid registering too many network callbacks that can lead to TooManyRequestsException being thrown since there is an application limit of 100. On API 30+ WorkManager will use a shared callback that can be registered via registerDefaultNetworkCallback() for which we can check individual requests using request.canBeSatisfiedBy(). For APIs 28-29 individual callbacks will still be registered but WorkManager will attempt to catch the TooManyRequestsException and report as if the constraint was not met.

Bug: 231499040
Test: NetworkRequestConstraintControllerTest
Change-Id: Iec82d038f0bc37a0130cfdc3e9722ece073c9fa2