Improve performance of InsetsRulers.
Bug: 409593614

Instead of watching every different ruler separately, any change
will now trigger the insets to update all together. They were
previously separated into rulers that are affected by IME and
rulers that aren't, but that distinction has been removed.

A new class was added to improve the performance of determining
the difference between ruler values set before and later, so that
appropriate LayoutNodes can be notified of the changes. It also
reduces repeated notification of request layout for nodes that
read multiple rulers.

Overall, the performance reduction on layout was between 55 us
and 107 us in one test and has been improved to 72 us. Therefore
the overhead has been reduced from 55 to 17 microseconds. This
performance issue should only be seen on first layout and when
the size of the ComposeView changes, so the degredation isn't
as bad as indicated in the tests.

Test: performance improvement for layout benchmarks
Change-Id: Ide0d3b6a83d8b88bb088596af97759c512ef9c52