Revert "Handle Time Broadcast using flow"
This reverts commit dcefb471cbd694964e7e1a783e65d94ef1182d43.

This change was intented to bring a performance increase in the
TimeText component by moving the registering of a BroadcastReceiver
off the main thread and onto Dispatchers.Default. However using
Dispatchers.Default does not work well with 1P apps that use TikTok,
as TikTok likes to manage its own threading, for further info see a
previous issue regarding this: b/389939842.

Bug: 411376602
Bug: 389939842
Test: N/A - Revert

Change-Id: I34d0292abaa91f7e69524818a130c9c5aa527cfa