Allow for frozen API changes to be baselined
Allows for added targets and top level declarations to be ignored on frozen APIs. Before this change we allowed ignoring changes but only incompatible changes, not compatible changes on frozen APIs.

Ideally we should report all compatible changes individually if shouldFreeze is true, but because we generally want to avoid changing frozen APIs and this is enough to unblock added new targets (which is reasonable).

Bug: 410849585
Test: Added new tests and validated changes in aosp/3500405 could be baselined
Change-Id: I294a8a119bc999be1808f683aba3353610739fa4