Configure destinationDirectory property for WasmJs compile tasks to avoid task invalidation
Without the changes, the wasmJsTestTestDevelopmentExecutableCompileSync and wasmJsTestTestProductionExecutableCompileSync were sharing the same `destinationDirectory`, and with the content they output being different, invalidate the cache each time the tasks are run together.

The configuration of `destinationDirectory` needs to be done after the browser DSL is complete as o/w, the properties would be overridden.

TESTED: Ran ./gradlew --ci buildOnServer createAllArchives checkExternalLicenses listTaskOutputs exportSboms -Pandroidx.enableComposeCompilerMetrics=true -Pandroidx.enableComposeCompilerReports=true --no-daemon on the compose runtime project and it passed

BUG: 375595154
Change-Id: I867f2a39d14e6057c44d918df489bc71073e0a50