Removed activeWindow and added waitForStable api
This pr removes activeWindow in UiAutomatorTestScope to avoid creating
confusion. Calling activeWindow right after startActivity or app returns
a window that is not the one of the app about to start. It's still
possible to access the active window via `activeWindowRoot().window`.

Also added `waitForStableInActiveWindow` to the `UiAutomatorTestScope`.
This is a direct replacement for `waitForIdle` that emphasizes the
monitored window is the active one.

Bug: 411307019
Test: ./gradlew test:uiautomator:uiautomator:cC
Relnote: Removed activeWindow and added waitForStable
Change-Id: I8b85177c9c40dfdcfb0b5b1ce09bbade34f86348