Introduce `fillableData` type for Autofill
This CL introduces a `FillableData` type that can be used by Autofill by multiple component types, rather than only text. The goal is to build semantic properties and actions on top of this type so that in the future, adding other component types such as list, toggle, and date, requires fewer APIs and less boilerplate code if they share this type in common.

See go/compose-fillable-data for more details and aosp/3581170 for a POC on how this type would be used.

Test: AndroidFillableDataTest
RelNote: TBD
Change-Id: I044b1e9f7f76ad5c5bf81a36203cb308d44fbaeb