Add RetainedEffect
RetainedEffect mirrors DisposableEffect for the retention lifecycle
instead of the remember lifecycle. This is especially useful for users
that want to execute effects on an object being retained if they either
can't or don't want to make their retained object implement
RetainObserver.

This effect is required to correctly implement the RetainedContentHost
API in a follow-up CL.

Bug: b/177562901
Test: EffectsTests
Relnote: """
    Introduces RetainedEffect, a side-effect API similar to
    DisposableEffect that follows the retention lifecycle instead of
    the composition lifecycle. This API is intended to be used for
    effects tied to the retention of another object that doesn't
    or can't implement RetainObserver.
    """
Change-Id: I1c61f7f644a99e2ca57bec4f617e443c2c0ac480