Add our own results bucket to ftl test runs
For running tests on FTL, we might be
able to use finer grained roles instead of the broad editor role, if we created our own storage bucket and
pass that as results storage.
https://firebase.google.com/docs/test-lab/android/iam-permissions-reference#glcoud-CLI-tests

When we don't pass this flag, it defaults to Firebase console generated bucket and accessing that requires the editor role.
https://cloud.google.com/sdk/gcloud/reference/firebase/test/android/run#--results-bucket

Change-Id: I38da18b1844b8699e1d324fd219e0794ae801391