Add title property to `DialogProperties`
This `dialogContentTitle` property added to `dialogProperties` can be set so that the default package announcement can be overridden. Before this CL, since the window title was not set, accessibility services defaulted to the package name when new dialogs were announced. Now this title can be overridden to be more specific and less redundant. See the bug for more details.

Test: dialogTest_windowTitleCustom()
Bug: 395546693
RelNote: Add in `dialogContentTitle` to `DialogProperties`, which sets the content window title.
Change-Id: I9fa5ab305cfd831cb92384ef05b0cc8af355fc06