Fix: Avoid additional onLoadDocumentSuccess callback
We're collecting fragment UI state using a repeatOnLifecycle API that is repeating in nature(cancels when out of target state and restarts when back in target state). This is causing document callbacks like onLoadDocumentSuccess to trigger again when going from background to foreground.

To fix this, start collecting using a one-shot api when fragment reaches the target state.

Bug: 409467284
Test: Manual
Change-Id: I8f18f3ca9d70d281ab3953731807ecce38414eb1