Fix issue with rememberNestedScrollInteropConnection not calling interop correctly.
We were closing the nested scroll cycle before it could finish, this would prevent onNestedFling from being called to the interop view. In this change I'm introducing a fix for this issue. A new ViewConfiguration property had to be introduced minimumFlingVelocity, to allow the correct axis
verification of the dispatching fling velocity.

Relnote: Introduced ViewConfiguration.minimumFlingVelocity to allow lower bound fling velocity control.

Fixes: 413704992
Test: Updated existing tests.
Change-Id: I11aab07292b0c3545149da8b4ac88bdc24154ae1