Set kotlinPlatforms in metalava project xml
The processing of the project xml file in lint no longer sets a module's platform to JVM if android=true is used.

This currently uses the default JVM platform, "JVM [1.8]" for all source sets. When metalava can handle non-JVM targets of KMP projects (b407737495/), this should be updated to actually compute the necessary platforms per source set.

Using JVM for the common module is currently causing b/322156458, but that is a bug on the lint side and will be addressed (this change maintains the existing buggy behavior).

Bug: 409337070
Test: ./gradlew checkApi, androidx_with_metalava target
Change-Id: Ic77a292f15856b1a545416ea5569ac1273dd5430