Remove hasValueForLayout check in Text, Spacer, ArcLine and DashedArcLine builders.
This CL removing the check for hasValueForLayout when a dynamic value is used. This will allow developers to not specify the valueForLayout properties when they use Text, Spacer, ArcLine and DashedArcLine.
If valueForLayout is not specified, the dimension will be calculated at runtime.

Test: Manual and unit tests
Bug: 413592806
Relnote: Remove hasValueForLayout check in Text, Spacer, ArcLine and DashedArcLine builders.
Change-Id: Ic52e8094e65c42b6db72c1279ca54b039dd5b182