Create a passthrough connection pool for drivers with internal connection pool.
Instead of special handling the AndroidSQLiteDriver and the SupportSQLiteDriver consolidate these two and any other driver that has an internal connection pool reported via the new API SQLiteDriver.hasConnectionPool to use a passthrough connection pool that does not do any connection management.

This also enables driver wrappers of AndroidSQLiteDriver as otherwise the instance check would fail leading Room to use the incorrect pool.

RoomSupportSQLite wrapper was also updated to provide compatibility for Room databases with drivers that have an internal pool, such as the AndroidSQLiteDriver.

Note that AndroidSQLiteConnection must be kept for backwards compatibility but was refactored to prepare statements of type SupportSQLiteStatement, while AndroidSQLiteStatement was merged with SupportSQLiteStatement and avoid repeated code. Moreover. in order to properly support the passthrough pool the driver handles transaction SQL statements directly.

Bug: 408010324
Test: Existing
Relnote: "Add an API to SQLiteDriver to report if internally it has a connection pool or not."
Change-Id: I52a5148dfeab4e576b613bb9d25f725d0c641a8e