Retrieve a11y services with `onViewAttachedToWindow`
After some investigation in b/400902455, we found that ComposeViews's accessibility information is not properly populated and distributed only while TalkBack is enabled when an app is in the background. When the app is hot-launched back into the foreground, TalkBack fails to focus.

This is because we enable a11y information to be distributed only when our two listeners (`stateChangeListener` and `touchExplorationStateChangeListener`) detect that a11y services are enabled. But when changes occur while the app containing the ComposeView is in the background, our listeners are removed since `onViewDetachedFromWindow` (and removing the listeners) occurs when we move the app to the background.

This change forces the Compose a11y delegate compat to retrieve enabled services when the ComposeView is reattached to the window, just in case any changes have happened while the app was in the background. I tested this with instructions linked in the bug; before these changes, TalkBack would not select the button upon opening the app, and after these changes, TalkBack does select the button upon opening the app.

Fixes: 400902455
Test: Manually dropped new artifacts into G3 to verify, see bug for more details
Change-Id: Ifd12a124dc56036a06136583a75fb142a76de3d3