Return true from AutoClosingSupportSQLiteDatabase.isOpen unless explicitly closed.
Since the database can be accessed outside of Room via roomDb.openHelper.writableDatabase it should not leak implementation details that it auto-closes. It is misleading to report the database is closed when it will re-open if interacted with it unless an explicit call to close() is done.

With this change RoomDatabase.isOpenInternal is updated to check directly with the AutoCloser which is used to avoid opening the database unnecessarily. This also means syncing triggers must be attempted even if the internal database is closed as otherwise manual start of transactions will lose invalidation notifications.

Test: AutoClosingRoomOpenHelperTest
Change-Id: I372a7b41e0c0783c578bc07de12f0af18ca2a9ef