DataStore Coroutine Context Sandwich 🥪.
This CL implements the coroutine context "sandwich" solution
proposed as a workaround for trace propagation issues encountered
when using a datastore.

The workaround combines context elements of the user (datastore)
coroutineContext and the caller's coroutineContext to ensure that
relevant CoroutineDispatcher.Key keys, essential for trace
propagation, are not lost when running datastore code in isolation
during read/write operations.

Bug: 370847635
Test: CoroutineContextSwitchTest.kt, StorageContextSwitchTest.kt
Change-Id: I6c214ba9327c72d4c9699559197faf22990806ef