Low Network Bandwidth Fix - Save Password
This continues the approach for the fix made in the 'Begin Sign In' case
to our save password operation. The focus was on the continuity of
the error flow. The exception messages now only rely directly from
GMS rather than contain our own indication of the exception moment.
Other than that, the flow should identically complete as before.

Bug: 415819849
Test: Tested on different network conditions

Change-Id: Ica9d85c7402f5b7ab82b98068330d99c28ffd119