Migrate KeepInView logic to pull based mechanism
The initial implementation of KeepInView logic in foundation had to use an observer to find the bounds of the focused child due to a limitation between `ui` and `foundation` layers. This limitation does not exist anymore and we can simply switch to a pull based mechanism where `ContentInViewNode` can just query its focused child to find its bounds.

Completely getting rid of this mechanism unfortunately also requires us to revert the solution that fixed reverse scrolling container problem. The new solution adjusts the container size and child placement for a single frame when a reverse scrolling viewport shrinks.

Bug: 392901531
Bug: 220030968
Test: existing tests for ContentInViewNode
Test: FocusTargetModifierNodeTest
Relnote: "Added `FocusTargetModifierNode.getFocusedRect()` that provides the focused area under this node's subtree, relative to the node's coordinates."
Relnote: "Added `ComposeFoundationFlags.isKeepInViewFocusObservationChangeEnabled` that enables the new way of keeping focused items in view for scrollable containers. If you run into a problem with the new approach, please file a bug and disable this flag."
Change-Id: I7a3dba71cc4bb2a193fa127bebdbe7b76b95110a