Edge Button refinement
Modified the algorithm for drawing the shape, removing
the morph into a pill-shaped button when it is very small,
now instead we make the circles slices at the top bigger
and the ellipsis arc at the bottom smaller, so the
transition is smoother.
The width of the button is now computed so that the minimum
distance betwen the button and the edge of the screen
is equal to the distance between the bottom of the button
and the bottom of the screen.
RelNote: "Improved the shape of the Edge Button"
Bug: 422364844
Test: Manual

Change-Id: I7721e8cff081c7b2a971ca0c0b6d05f1ffa690ea