Remove old transition APIs
We currently have 2 ways to call transitions, we should start with 1 and
we can add another later if needed. So we will only take transitionSpec
for right now.

With the NavDisplay that takes the transitionSpec, we also need to take
a SizeTransform separate. For the navigation case, we want the
SizeTransform to be null by default so that animations don't clip in
unexpected ways. For ContentTransform, the SizeTransform is non-null by
default so we can't tell when developers have a custom SizeTransfrom
inside of their ContentTransform, or they just didn't set the
SizeTransform at all.

RelNote: N/A
Test: modified all transition to use specs
Change-Id: Icea8ce731579bfbf0ccf82142ce7e3d760dd8380