Provide the NavKey marker class
Adding a new NavKey class that is meant to be used as a marker to
indiciate that an extending class or object is marked as Serializable
and can be saved. This comes as a result of the fact that classes marked
as @Serializable in Kotlin aren't actually of the same type so there is
no way to uniformly save and store those classes.

With the addition of the NavKey, we can now add a rememberNavBackStack
that is capable of automatically saving components that extend the
NavKey and a rememberMutableStateListOf that can save all types.

RelNote: N/A
Bug: 403272366
Bug: 403272377
Test: moved test using rememberSaveable to use rememberNavBackStack and
in sample app

Change-Id: I3108efc5e565e609fb658dcee96427a05db224d2