Fix Room not exporting schemas in native compilation
The issue is that in non-JVM the annotation default values are missing in KSP2 and room-compiler should detect it and use the annotation default value but for boolean values the wrong expression was being used, not using the elvis operator and using an equality check. When equality is used, if the value is missing then it will return 'false' ('null' is not equal to 'true') which works for the many default annotation values that are 'false', but not `exportSchema` whose default value is 'true'. This change updates all boolean annotation checks.

Also re-enable RoomKmpGradlePluginTest which would have caught this but where disabled due to an issue that doesn't seem to be present anymore, likely after updating Kotlin toolchain. The test project was also updated to Kotlin 2.1 along with minor changes to adhere to 2.0 expect / actual rules.

Bug: 374360882
Bug: 416549580
Test: RoomKmpGradlePluginTest
Change-Id: Ib4b4f28c0f1069323377a2398b45f4b9188d8a57