Resolve Compose Composition Issues in SpatialDialog and ElevatedPanel
The use of the view to determine the size of the dialog was unnecessary and lead to incorrect sizing. Instead, we constrain it with empty constraints and allow it to measure itself to size it.

Added a TODO to investigate if `ElevatedPanel` still needs to be `alpha` dependent on Pose, as this may not be needed.

Test: Manually on the SpatialComposeApp on device on current sys img

Fixes: 401619909

Relnote: Fixed issue where SpatialDialog would flash when being rendered

Change-Id: Ife73c018a39141e0d1d524709d4577b1372c09e3