Stable version of LazyLayoutPrefetchState
In this CL we're introducing a stable version of LazyLayoutPrefetchState with basic prefetch support.  This unblocks the creation of custom LazyLayouts with basic prefetch support. For compatibility we're keeping an additional experimental constructor for nested prefetch and scheduler customizations.

Relnote: Stabilized an empty constructor for LazyLayoutPrefetchState and it's scheduling method schedulePrecomposition and schedulePrecompositionAndPremeasure.

Bug: 252853717
Test: Existing tests

Change-Id: I4362f3417ebe652d967463a474e4ba799967930d