Maintain the application spatial compose scene in memory.
This maintains application state when the application is in the background.

See design doc at go/compose-xr:application-state-changes

https://screencast.googleplex.com/cast/NTE4MjAzMTM5MzcxODI3MnwxMTg2OTUwZC1mZA
https://screencast.googleplex.com/cast/NjQ4NTU1NzE4OTI3OTc0NHxjMGQ0OTdmMS02YQ

Relnote: Subspaces and Orbiters will now retain their internal state in home space and when the app is in the background. In home space mode, Subspace will still set up its scene in preparation for the switch to full space mode.
Bug: 416037751
Test: Ran SpatialComposeApp locally. Ran unit tests.
Change-Id: I403176b040bbeb62ea0c4ab098992a6232da1c1d