Fix issue with inactive NestedScrollNode coroutine Scope.
Nested scroll nodes will set in the dispatcher the outer-most coroutine scope so events started at a child nested scroll node can outlive them. The system falls back to the dispatcher node when the parent node is not set. We're updating this to also fallback in case the parent's scope is no longer active. There are cases where the parent scope is no longer active so we don't want to use that scope.

Relnote: Introduce flag isNestedScrollDispatcherNodeFixEnabled.

Fixes: 411603219
Test: Added new test.
Change-Id: I0d24ad6c222a319c9a9fb0356309a3b99fa15418