Revert^2 "Add toColorLong() and fromColorLong() APIs to Color"
This reverts commit bd533c47606e02afcbb94d922d5cc0b5a879b370.

Reason for revert: fix for error

Add toColorLong() and fromColorLong() APIs to Color

Fixes: 343736590

Relnote: "Compose 64-bit color values are not directly comparable
to Android ColorLongs because the color space IDs are out of order
for some color spaces. To convert to and from Android color spaces,
two new APIs are added: toColorLong() and fromColorLong()."

This also changes usages of Long values from Colors.

Test: new tests
Change-Id: Iffe5ca7d8152b10402cedd887529f3247a90fa73