Fix LifecycleCameraProviderImpl leak
LifecycleCameraProviderImpl holds a list of
LifecycleCameraRepository.Key in order to only remove specific cameras
when shutodown. This list will not be cleared unless shutdown is called.

LifecycleCameraRepository.Key as a identifier doesn't necessarily need
to own the lifecycleOwner. This CL changes it to only store the hashCode
to prevent the leackage.

Bug: b/417070509
Test: ./gradlew camera:camera-lifecycle:build
Change-Id: I2c4ad99d934d4471a8df4dc592b28c1dd9f33e81