Make sure ComposeView children are attached before creating Composition
This is a fix to a ticket which exposes race condition between
activities destroy/create callbacks and AndroidComposeView reacting to
those. The consequence of that issue is a black screen without any
composables.

The problem is seen when ComposeView is created once and reused between
activities by removeView() from previous activity and addView() to the
new one. If this happens very quickly, by running commands:

adb shell input keyevent 4
adb shell am start -a android.intent.action.MAIN -c
  android.intent.category.HOME androidx.wear.compose.integration.demos/
  androidx.wear.compose.integration.demos.DemoActivity

Then onDetachedToWindow/onAttachedToWindow can happen in an unexpected
order. The issue is because:
1. User pressed back, which starts current activity end of life cycle
2. New activity starts with adb shell am start, which:
   a) Removes view from the old activity by callling removeView(),
      which goes down to AndroidComposeView::onDetachedFromWindow().
      At this point AndroidComposeView::viewTreeOwner will point at
      activity which is currently being destroyed
   b) Add ComposeView to FrameLayout of the new activity
   c) AbstractComposeView receives onAttachedToWindow callback, which
      calls down to AndroidComposeView.setOnViewTreeOwnersAvailable()
      where it registeres lifecycle observers to the current
      viewTreeOwner (which is an old activity)
   d) Old activity closes which triggers newly registered lifecycle to
      call WrappedComposition.dispose()
   e) All composables are removed from the screen

Fixes: 415046577
Test: Two new tests composeViewProperlyChangesLifecycleOwner and
      composeViewNotDestroyedWhenOldLifecycleOwnerIsDestroyed
Relnote: N/A
Change-Id: I2d5d06196c0019d7b67c71d692cba9af654985b2