Add overlay scene support and DialogSceneStrategy
Provide a subclass of Scene called OverlayScene
that allows a SceneStrategy to mark its scene as an
'overlay' that sits above another Scene.

A naive implementation in DialogSceneStrategy is
the simplest version of this, wrapping the content
of the entry in a Dialog.

Note the TODOs where moving content from a dialog
to another scene will be handled in a follow up CL.

Test: updated NavDisplayTest
Relnote: "`SceneStrategy` implementations can now
choose to return an instance of `OverlayScene` which
enables dialogs, bottom sheets, and other types of
overlays. A `DialogSceneStrategy` can be added to your
`NavDisplay` to enable entries to be displayed within
a composable `Dialog`."
BUG: 205482229

Change-Id: Ifa3bba4c6cbfed3fe619915dcf1366df51481579