Only track MR2 session records in MR2ProviderServiceAdapter
The current implementation of the two methods in
MR2ProviderServiceAdapter `notifyRouteControllerAdded` and
`notifyRouteControllerRemoved` track route controllers created
by registered media route providers. So the mSessionRecords
contains a lot of sessions with requestId = 0 and non-MR2
sessions (i.e. flags | SESSION_FLAG_MR2 == 0). Those
sessionRecords will be updated to the MR2 framework which
caused a lot of warning messages with unknown requests (i.e.
requestId = 0). Also, those mSessionRecords are added to
the mSessionInfos in the MR2 framework. When we assume that
the last RoutingSessionInfo is the last item in the
mSessionInfos, we easily get a wrong RoutingSessionInfo with
requestId = 0. In this case, the following transfer will
fail and is stuck in the failed state until we restart the
phone. Removing those SessionRecords with requestId = 0
could improve the quality of MR2.

Bug: b/417302861
Test: verify with the sample app and casting with multiple apps at the same time
Change-Id: Ibc185326c5d74ce217fc83699cc79e5ba65c58bf