Add Frame End Scheduling APIs
This commit adds basic in-between frame scheduling to implement a public
`awaitOrScheduleNextFrameEnd()` method. This is intended to be used by
RetainScopes to end retention after composition settles.

This behavior shares a lot in common with `BroadcastFrameClock`, so this
commit also factors out most of the awaiters logic into a reusable
`AwaiterQueue` class.

Test: RecomposerFrameEndSchedulingTests
Relnote: """
    Added an `awaitOrScheduleNextCompositionEnd()` API, which issues a
    callback when the Recomposer finishes composing the current frame.
    If the Recomposer isn't currently composing a frame, this API
    suspends until the next frame's composition finishes and schedules
    a new frame to be composed.
    """

Change-Id: Ib873cd7d64e63f6337232634faf04ee9e711f7fd