Add missing transaction context switch in passthrough pool.
In Ic3055d9e6c284600cdde532a033b8313645a1660 the passthrough pool added switching to the transaction context if available for Android when in compatibility mode when starting a transaction. But it missed, similar to DBUtil to switch to the transaction context when using the statement via the pool APIs.

Bug: 415006268
Test: CompatibilityModeTest
Change-Id: I85b768aef7c693dd66562a8424973fa7dfd795bc