Add transitionSpec specific for PredictiveBack
Adding a predictivePopTransitionSpec to only be when there is a
predictive back being handled. This will allow users to separate the
predictive back case out from the regular pop case.

The the predictivePopTransitionSpec on the NavDisplay defaults so the
same as the popTransitionSpec while the one for the NavEntry metadata
defaults to null.

RelNote: N/A
Test: changed predictive back tests to use new spec and used it in
sample app

Change-Id: Icbdea45ee7cf1079a79d73c67ee6c1b9648bbd9e