Fix handling of the reusing state for pausable composition
Fixes how reusing state is handled in the composer for pausable
composition. The previous code assume a restart scope will be
restarted directly and the setting of the reusing state was in
`recomposeToGroupEnd()`. This ignored the case were a parent
scope was also invalidated and called the paused scope directly.
This moves the tracking of the reusing state to
`startRestartGroup()` and `endRestartGroup()`.

Fixes: b/404058957
Test: ./gradlew :compose:r:r:i-t:conAIT
Relnote: "Fix for for reusing state handling in pausable
composition"

Change-Id: Ife96e7e011daa3644cdda2f112bbe7bbedb49ae4