change incoming call flow - add call in parallel to notification
Previously, when a new incoming call request came in to the broadcast
receiver, only an incoming call style notification would be made. The
call would only be added to telecom when the answer button was clicked.
This is not the call flow we want to show as an example.

Instead, the call should be added to telecom in parallel to creating
the call-style notification.

Fixes: 417547684
Test: manual
Change-Id: I58a46159d46813e03da78698d2f40ea418c0bd12